# zhipin-kratos/api/zhipin/requirement_item_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.0
// source: zhipin/requirement_item.proto

package zhipin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationRequirementItemServiceListRequirementItems = "/api.zhipin.RequirementItemService/ListRequirementItems"
const OperationRequirementItemServiceSetRequirementItems = "/api.zhipin.RequirementItemService/SetRequirementItems"

type RequirementItemServiceHTTPServer interface {
	ListRequirementItems(context.Context, *ListRequirementItemsReq) (*ListRequirementItemsResp, error)
	SetRequirementItems(context.Context, *SetRequirementItemsReq) (*SetRequirementItemsResp, error)
}

func RegisterRequirementItemServiceHTTPServer(s *http.Server, srv RequirementItemServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/position/{position_id}/requirement-items", _RequirementItemService_SetRequirementItems0_HTTP_Handler(srv))
	r.GET("/api/position/{position_id}/requirement-items", _RequirementItemService_ListRequirementItems0_HTTP_Handler(srv))
}

func _RequirementItemService_SetRequirementItems0_HTTP_Handler(srv RequirementItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SetRequirementItemsReq
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRequirementItemServiceSetRequirementItems)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SetRequirementItems(ctx, req.(*SetRequirementItemsReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SetRequirementItemsResp)
		return ctx.Result(200, reply)
	}
}

func _RequirementItemService_ListRequirementItems0_HTTP_Handler(srv RequirementItemServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListRequirementItemsReq
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationRequirementItemServiceListRequirementItems)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListRequirementItems(ctx, req.(*ListRequirementItemsReq))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListRequirementItemsResp)
		return ctx.Result(200, reply)
	}
}

type RequirementItemServiceHTTPClient interface {
	ListRequirementItems(ctx context.Context, req *ListRequirementItemsReq, opts ...http.CallOption) (rsp *ListRequirementItemsResp, err error)
	SetRequirementItems(ctx context.Context, req *SetRequirementItemsReq, opts ...http.CallOption) (rsp *SetRequirementItemsResp, err error)
}

type RequirementItemServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewRequirementItemServiceHTTPClient(client *http.Client) RequirementItemServiceHTTPClient {
	return &RequirementItemServiceHTTPClientImpl{client}
}

func (c *RequirementItemServiceHTTPClientImpl) ListRequirementItems(ctx context.Context, in *ListRequirementItemsReq, opts ...http.CallOption) (*ListRequirementItemsResp, error) {
	var out ListRequirementItemsResp
	pattern := "/api/position/{position_id}/requirement-items"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationRequirementItemServiceListRequirementItems))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *RequirementItemServiceHTTPClientImpl) SetRequirementItems(ctx context.Context, in *SetRequirementItemsReq, opts ...http.CallOption) (*SetRequirementItemsResp, error) {
	var out SetRequirementItemsResp
	pattern := "/api/position/{position_id}/requirement-items"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationRequirementItemServiceSetRequirementItems))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}