# zhipin-kratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/position/{id}/requirements:
        put:
            tags:
                - PositionService
            operationId: PositionService_UpdateRequirements
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.UpdateRequirementsReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.PositionResp'
...
    /api/position/{positionId}/requirement-items:
        get:
            tags:
                - RequirementItemService
            operationId: RequirementItemService_ListRequirementItems
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.ListRequirementItemsResp'
        post:
            tags:
                - RequirementItemService
            operationId: RequirementItemService_SetRequirementItems
            parameters:
                - name: positionId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.zhipin.SetRequirementItemsReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.zhipin.SetRequirementItemsResp'
...
components:
    schemas:
...
        api.zhipin.ListRequirementItemsResp:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.zhipin.RequirementItemResp'
...
        api.zhipin.PositionResp:
            type: object
            properties:
                id:
                    type: string
                jobId:
                    type: string
                title:
                    type: string
                company:
                    type: string
                salaryRange:
                    type: string
                salaryMin:
                    type: integer
                    format: int32
                salaryMax:
                    type: integer
                    format: int32
                city:
                    type: string
                link:
                    type: string
                recruiter:
                    type: string
                encBossId:
                    type: string
                isHunter:
                    type: boolean
                status:
                    type: integer
                    format: int32
                skipReason:
                    type: string
                matchRate:
                    type: integer
                    format: int32
                notes:
                    type: string
                lastCommAt:
                    type: string
                lastCommDir:
                    type: integer
                    format: int32
                lastResume:
                    type: string
                duties:
                    type: string
                requirements:
                    type: string
...
        api.zhipin.RequirementItemInput:
            type: object
            properties:
                requirement:
                    type: string
                matchStatus:
                    type: integer
                    format: int32
                resumePoint:
                    type: string
                remark:
                    type: string
                sortIndex:
                    type: integer
                    format: int32
        api.zhipin.RequirementItemResp:
            type: object
            properties:
                id:
                    type: string
                positionId:
                    type: string
                requirement:
                    type: string
                matchStatus:
                    type: integer
                    format: int32
                resumePoint:
                    type: string
                remark:
                    type: string
                sortIndex:
                    type: integer
                    format: int32
        api.zhipin.SetRequirementItemsReq:
            type: object
            properties:
                positionId:
                    type: string
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.zhipin.RequirementItemInput'
        api.zhipin.SetRequirementItemsResp:
            type: object
            properties:
                count:
                    type: integer
                    format: int32
...
        api.zhipin.UpdateRequirementsReq:
            type: object
            properties:
                id:
                    type: string
                requirements:
                    type: string
...